qr->photo 이동
